YARN-10408. Extract MockQueueHierarchyBuilder to a separate class. Contributed by Gergely Pollak
